requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
html2text==2024.2.26
playwright==1.49.1
pydantic==2.10.4
slowapi==0.1.9
python-dotenv==1.0.1
